RickAndMortyClient/backend/requirements.txt
Django>=5.2.4
djangorestframework
requests
psycopg2-binary
